Export Bot API types
