+++ package.json
@@ -40,6 +40,10 @@
     ],
     "main": "./out/mod.js",
     "types": "./out/mod.d.ts",
+    "exports": {
+        ".": "./out/mod.js",
+        "./types": "./out/types.js"
+    },
     "keywords": [
         "telegram",
         "bot",
