+++ src/bot.ts
@@ -7,11 +7,8 @@ import {
     type WebhookReplyEnvelope,
 } from "./core/client.ts";
 import { GrammyError, HttpError } from "./core/error.ts";
-import {
-    debug as d,
-    type Update,
-    type UserFromGetMe,
-} from "./platform.deno.ts";
+import { debug as d } from "./platform.deno.ts";
+import { type Update, type UserFromGetMe } from "./types.ts";
 const debug = d("grammy:bot");
 const debugErr = d("grammy:error");
