+++ src/core/payload.ts
@@ -1,4 +1,5 @@
-import { InputFile, itrToStream, toRaw } from "../platform.deno.ts";
+import { InputFile } from "../types.ts";
+import { itrToStream, toRaw } from "../platform.deno.ts";
 
 // === Payload types (JSON vs. form data)
 /**
