+++ src/types.node.ts
@@ -1,0 +1,139 @@
+// === Needed imports
+import { type InputFileProxy } from "@grammyjs/types";
+import { debug as d, toRaw } from "./platform.node";
+import { basename } from "path";
+import fetch from "node-fetch";
+import { URL } from "url";
+import { createReadStream, type ReadStream } from "fs";
+
+const debug = d("grammy:warn");
+
+export * from "@grammyjs/types";
+
+// === Export all API types
+/** Something that looks like a URL. */
+interface URLLike {
+    /**
+     * Identifier of the resouce. Must be in a format that can be parsed by the
+     * URL constructor.
+     */
+    url: string;
+}
+
+// === InputFile handling and File augmenting
+/**
+ * An `InputFile` wraps a number of different sources for [sending
+ * files](https://grammy.dev/guide/files.html#uploading-your-own-file).
+ *
+ * It corresponds to the `InputFile` type in the [Telegram Bot API
+ * Reference](https://core.telegram.org/bots/api#inputfile).
+ */
+export class InputFile {
+    private consumed = false;
+    private readonly fileData: ConstructorParameters<typeof InputFile>[0];
+    /**
+     * Optional name of the constructed `InputFile` instance.
+     *
+     * Check out the
+     * [documenation](https://grammy.dev/guide/files.html#uploading-your-own-file)
+     * on sending files with `InputFile`.
+     */
+    public readonly filename?: string;
+    /**
+     * Constructs an `InputFile` that can be used in the API to send files.
+     *
+     * @param file A path to a local file or a `Buffer` or a `fs.ReadStream` that specifies the file data
+     * @param filename Optional name of the file
+     */
+    constructor(
+        file:
+            | string
+            | URL
+            | URLLike
+            | Uint8Array
+            | ReadStream
+            | Iterable<Uint8Array>
+            | AsyncIterable<Uint8Array>,
+        filename?: string,
+    ) {
+        this.fileData = file;
+        filename ??= this.guessFilename(file);
+        this.filename = filename;
+        if (
+            typeof file === "string" &&
+            (file.startsWith("http:") || file.startsWith("https:"))
+        ) {
+            debug(
+                `InputFile received the local file path '${file}' that looks like a URL. Is this a mistake?`,
+            );
+        }
+    }
+    private guessFilename(
+        file: ConstructorParameters<typeof InputFile>[0],
+    ): string | undefined {
+        if (typeof file === "string") return basename(file);
+        if (typeof file !== "object") return undefined;
+        if ("url" in file) return basename(file.url);
+        if (!(file instanceof URL)) return undefined;
+        return basename(file.pathname) || basename(file.hostname);
+    }
+    [toRaw](): Uint8Array | Iterable<Uint8Array> | AsyncIterable<Uint8Array> {
+        if (this.consumed) {
+            throw new Error("Cannot reuse InputFile data source!");
+        }
+        const data = this.fileData;
+        // Handle local files
+        if (typeof data === "string") return createReadStream(data);
+        // Handle URLs and URLLike objects
+        if (data instanceof URL) {
+            return data.protocol === "file" // node-fetch does not support file URLs
+                ? createReadStream(data.pathname)
+                : fetchFile(data);
+        }
+        if ("url" in data) return fetchFile(data.url);
+        // Mark streams and iterators as consumed
+        if (!(data instanceof Uint8Array)) this.consumed = true;
+        // Return buffers and byte streams as-is
+        return data;
+    }
+}
+
+async function* fetchFile(url: string | URL): AsyncIterable<Uint8Array> {
+    const { body } = await fetch(url);
+    for await (const chunk of body) {
+        if (typeof chunk === "string") {
+            throw new Error(
+                `Could not transfer file, received string data instead of bytes from '${url}'`,
+            );
+        }
+        yield chunk;
+    }
+}
+
+// === Export InputFile types
+type GrammyTypes = InputFileProxy<InputFile>;
+
+/** Wrapper type to bundle all methods of the Telegram API */
+export type ApiMethods = GrammyTypes["Telegram"];
+
+/** Utility type providing the argument type for the given method name or `{}` if the method does not take any parameters */
+export type Opts<M extends keyof GrammyTypes["Telegram"]> =
+    GrammyTypes["Opts"][M];
+
+/** This object represents the content of a media message to be sent. It should be one of
+- InputMediaAnimation
+- InputMediaDocument
+- InputMediaAudio
+- InputMediaPhoto
+- InputMediaVideo */
+export type InputMedia = GrammyTypes["InputMedia"];
+/** Represents a photo to be sent. */
+export type InputMediaPhoto = GrammyTypes["InputMediaPhoto"];
+/** Represents a video to be sent. */
+export type InputMediaVideo = GrammyTypes["InputMediaVideo"];
+/** Represents an animation file (GIF or H.264/MPEG-4 AVC video without sound) to be sent. */
+export type InputMediaAnimation = GrammyTypes["InputMediaAnimation"];
+/** Represents an audio file to be treated as music to be sent. */
+export type InputMediaAudio = GrammyTypes["InputMediaAudio"];
+/** Represents a general file to be sent. */
+export type InputMediaDocument = GrammyTypes["InputMediaDocument"];
